requirements.txt
imageio==2.25.0
imutils==0.5.4
kaleido>=0.2.1
matplotlib==3.6.3
metpy==1.4.0
netcdf4==1.6.2
numpy==1.24.2
opencv-python>=4.7.0.72
pandas==1.5.2
pint==0.20.1
plotly==5.13.0
pytest>=7.2.1
pyyaml==6.0
scikit-image==0.19.3
scipy>=1.11.1 
xarray==2023.1.0
eofs==1.4.0
#cartopy==0.21.1
scikit-learn==1.5.0
